Implement 'rerun tests' functionality

+++ build-packages/build.sh
@@ -81,41 +81,41 @@ do
   fi
 done
 
-cd $project_path
-git submodule update --init
-git remote rm origin
-git checkout $commit_hash
+if [ "$rerun_tests" != 'true' ] ; then
+  cd $project_path
+  git submodule update --init
+  git remote rm origin
+  git checkout $commit_hash
 
-# Downloads extra files by .abf.yml
-ruby $rpm_build_script_path/../abf_yml.rb -p $project_path
+  # Downloads extra files by .abf.yml
+  ruby $rpm_build_script_path/../abf_yml.rb -p $project_path
 
-# Check count of *.spec files (should be one)
-x=`ls -1 | grep '.spec$' | wc -l | sed 's/^ *//' | sed 's/ *$//'`
-spec_name=`ls -1 | grep '.spec$'`
-if [ $x -eq '0' ] ; then
-  echo '--> There are no spec files in repository.'
-  exit 1
-else
-  if [ $x -ne '1' ] ; then
-    echo '--> There are more than one spec file in repository.'
+  # Check count of *.spec files (should be one)
+  x=`ls -1 | grep '.spec$' | wc -l | sed 's/^ *//' | sed 's/ *$//'`
+  spec_name=`ls -1 | grep '.spec$'`
+  if [ $x -eq '0' ] ; then
+    echo '--> There are no spec files in repository.'
     exit 1
+  else
+    if [ $x -ne '1' ] ; then
+      echo '--> There are more than one spec file in repository.'
+      exit 1
+    fi
   fi
+
+  # build changelog (limited to ~10 for reasonable changelog size)
+  sed -i '/%changelog/,$d' $spec_name
+  echo >> $spec_name
+  echo '%changelog' >> $spec_name
+  changelog_log=$results_path/changelog.log
+  echo "python $rpm_build_script_path/build-changelog.py -b 5 -e $commit_hash -n $spec_name >> $changelog_log"
+  python $rpm_build_script_path/build-changelog.py -b 5 -e $commit_hash -n $spec_name >> $changelog_log
+  echo "cat $changelog_log >> $spec_name"
+  cat $changelog_log >> $spec_name
+
+  # Remove .git folder
+  rm -rf $project_path/.git
 fi
-
-
-# build changelog (limited to ~10 for reasonable changelog size)
-sed -i '/%changelog/,$d' $spec_name
-echo >> $spec_name
-echo '%changelog' >> $spec_name
-changelog_log=$results_path/changelog.log
-echo "python $rpm_build_script_path/build-changelog.py -b 5 -e $commit_hash -n $spec_name >> $changelog_log"
-python $rpm_build_script_path/build-changelog.py -b 5 -e $commit_hash -n $spec_name >> $changelog_log
-echo "cat $changelog_log >> $spec_name"
-cat $changelog_log >> $spec_name
-
-
-# Remove .git folder
-rm -rf $project_path/.git
 
 if [[ "$platform_arch" == "armv7l" || "$platform_arch" == "armv7hl" ]]; then
   cd $rpm_build_script_path
@@ -202,6 +202,31 @@ EXTRA_CFG_OPTIONS="$extra_cfg_options" \
 
 r=`cat $config_dir/default.cfg | grep "config_opts\['root']" | awk '{ print $3 }' | sed "s/'//g"`
 chroot_path=$tmpfs_path/$r
+
+# Rerun tests
+if [ "$rerun_tests" == 'true' ] ; then
+  export RERUN_TESTS='false' \
+       PACKAGES=${packages} \
+       results_path=$results_path \
+       tmpfs_path=$tmpfs_path \
+       rpm_path=$rpm_path \
+       chroot_path=$chroot_path \
+       src_rpm_path=$src_rpm_path \
+       rpm_build_script_path=$rpm_build_script_path \
+       use_extra_tests=$use_extra_tests \
+       platform_name=$platform_name \
+       platform_arch=$platform_arch
+
+  /bin/bash $rpm_build_script_path/tests.sh
+  # Save exit code
+  rc=$?
+  if [ ${rc} != 0 ] ; then
+    echo '--> Test failed, see: tests.log'
+    exit 5
+  fi
+  exit 0
+fi
+
 # Download tarball with existing chroot, if any
 # The tarball should contain 'root' folder which will be unpacked
 # to the directory used by mock-urpm
@@ -297,17 +322,18 @@ rpm -qa --queryformat "%{name}-%{version}-%{release}.%{arch}.%{disttag}%{distepo
 test_code=0
 if [ $rc == 0 ]
 then
-  export results_path=$results_path \
+  export RERUN_TESTS='false' \
+       results_path=$results_path \
        tmpfs_path=$tmpfs_path \
        rpm_path=$rpm_path \
        chroot_path=$chroot_path \
        src_rpm_path=$src_rpm_path \
-       use_extra_tests=$use_extra_tests \
        rpm_build_script_path=$rpm_build_script_path \
+       use_extra_tests=$use_extra_tests \
        platform_name=$platform_name \
        platform_arch=$platform_arch
 
-  /bin/bash $rpm_build_script_path/test.sh
+  /bin/bash $rpm_build_script_path/tests.sh
   test_code=$?
 fi
